Rename RuleTypes to ruleTypes
To make sure all file names are camel case
