+++ src/types/index.ts
@@ -1,3 +1,3 @@
-export { SizeRule, FormatsRule, DurationRule, CustomRule, Rule, FactoryRule, CustomRuleValidator } from './RuleTypes'
+export { SizeRule, FormatsRule, DurationRule, CustomRule, Rule, FactoryRule, CustomRuleValidator } from './ruleTypes'
 export { UFSFile, FileType } from './fileTypes'
 export { UFSError, UFSCustomError } from './errorTypes'
